Refactor directory stucture into modalities for text and vision
